Move parallel and sequence builders to top-level classes
