+++ commandsv3/src/main/java/org/wpilib/commands3/ParallelGroup.java
@@ -6,13 +6,10 @@ package org.wpilib.commands3;
 
 import static edu.wpi.first.util.ErrorMessages.requireNonNullParam;
 
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.Set;
-import java.util.function.BooleanSupplier;
-import java.util.stream.Collectors;
 
 /**
  * A type of command that runs multiple other commands in parallel. The group will end after all
@@ -98,7 +95,7 @@ public class ParallelGroup implements Command {
    * @param commands the commands to run in parallel
    * @return the group
    */
-  public static Builder race(Command... commands) {
+  public static ParallelGroupBuilder race(Command... commands) {
     return builder().optional(commands);
   }
 
@@ -109,7 +106,7 @@ public class ParallelGroup implements Command {
    * @param commands the commands to run in parallel
    * @return the group
    */
-  public static Builder all(Command... commands) {
+  public static ParallelGroupBuilder all(Command... commands) {
     return builder().requiring(commands);
   }
 
@@ -153,126 +150,7 @@ public class ParallelGroup implements Command {
     return "ParallelGroup[name=" + name + "]";
   }
 
-  public static Builder builder() {
-    return new Builder();
-  }
-
-  public static class Builder {
-    private final Set<Command> commands = new LinkedHashSet<>();
-    private final Set<Command> requiredCommands = new LinkedHashSet<>();
-    private BooleanSupplier endCondition = null;
-
-    /**
-     * Adds one or more optional commands to the group. They will not be required to complete for
-     * the parallel group to exit, and will be canceled once all required commands have finished.
-     *
-     * @param commands The optional commands to add to the group
-     * @return The builder object, for chaining
-     */
-    public Builder optional(Command... commands) {
-      this.commands.addAll(Arrays.asList(commands));
-      return this;
-    }
-
-    /**
-     * Adds one or more required commands to the group. All required commands will need to complete
-     * for the group to exit.
-     *
-     * @param commands The required commands to add to the group
-     * @return The builder object, for chaining
-     */
-    public Builder requiring(Command... commands) {
-      requiredCommands.addAll(Arrays.asList(commands));
-      this.commands.addAll(requiredCommands);
-      return this;
-    }
-
-    /**
-     * Forces the group to be a pure race, where the group will finish after the first command in
-     * the group completes. All other commands in the group will be canceled.
-     *
-     * @return The builder object, for chaining
-     */
-    public Builder racing() {
-      requiredCommands.clear();
-      return this;
-    }
-
-    /**
-     * Forces the group to require all its commands to complete, overriding any configured race or
-     * deadline behaviors. The group will only exit once every command has completed.
-     *
-     * @return The builder object, for chaining
-     */
-    public Builder requireAll() {
-      requiredCommands.clear();
-      requiredCommands.addAll(commands);
-      return this;
-    }
-
-    /**
-     * Adds an end condition to the command group. If this condition is met before all required
-     * commands have completed, the group will exit early. If multiple end conditions are added
-     * (e.g. {@code .until(() -> conditionA()).until(() -> conditionB())}), then the last end
-     * condition added will be used and any previously configured condition will be overridden.
-     *
-     * @param condition The end condition for the group
-     * @return The builder object, for chaining
-     */
-    public Builder until(BooleanSupplier condition) {
-      endCondition = condition;
-      return this;
-    }
-
-    /**
-     * Creates the group, using the provided named. The group will require everything that the
-     * commands in the group require, and will have a priority level equal to the highest priority
-     * among those commands.
-     *
-     * @param name The name of the parallel group
-     * @return The built group
-     */
-    public ParallelGroup named(String name) {
-      var group = new ParallelGroup(name, commands, requiredCommands);
-      if (endCondition == null) {
-        // No custom end condition, return the group as is
-        return group;
-      }
-
-      // We have a custom end condition, so we need to wrap the group in a race
-      return builder()
-          .optional(group, Command.waitingFor(endCondition).named("Until Condition"))
-          .named(name);
-    }
-
-    /**
-     * Creates the group, giving it a name based on the commands within the group.
-     *
-     * @return The built group
-     */
-    public ParallelGroup withAutomaticName() {
-      // eg "(CommandA & CommandB & CommandC)"
-      String required =
-          requiredCommands.stream().map(Command::name).collect(Collectors.joining(" & ", "(", ")"));
-
-      var optionalCommands = new LinkedHashSet<>(commands);
-      optionalCommands.removeAll(requiredCommands);
-      // eg "(CommandA | CommandB | CommandC)"
-      String optional =
-          optionalCommands.stream().map(Command::name).collect(Collectors.joining(" | ", "(", ")"));
-
-      if (requiredCommands.isEmpty()) {
-        // No required commands, pure race
-        return named(optional);
-      } else if (optionalCommands.isEmpty()) {
-        // Everything required
-        return named(required);
-      } else {
-        // Some form of deadline
-        // eg "[(CommandA & CommandB) * (CommandX | CommandY | ...)]"
-        String name = "[%s * %s]".formatted(required, optional);
-        return named(name);
-      }
-    }
+  public static ParallelGroupBuilder builder() {
+    return new ParallelGroupBuilder();
   }
 }
